rename UserControler to BaseAPIControler and create 2 funtions sendSuccess() and sendError()
